Extract important queries to postgres functions
The motivating use case for extracting these functions is that it
gives more control to the developer using this package on how to
deploy this to their infrastructure.

For instance, some people may be interested in a different table
structure for `payloads`, like having it partitioned. With this
change, developers are free to change the implementation of the
postgres functions as they see fit as long as they implement an
equivalent logic and return the same types.